+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 - adding HTTP-Proxy (this is a backward-compatible change)
+- adding `seedfarmer list allmoduledata` to fetch all the metadata in a deployment in a single call
+
 
 ### Changes